requirements.txt
amqp==5.1.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-extensions==3.1.5
django-import-export==2.7.1
django-timezone-field==4.2.3
et-xmlfile==1.1.0
idna==3.3
kombu==5.2.4
MarkupPy==1.14
mysqlclient==2.1.0
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==2.67.0
tablib==3.2.0
urllib3==1.26.8
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
